multi-shot-scanner/backend/requirements.txt
# File: backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
anthropic==0.34.2
python-dotenv==1.0.0
Pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
werkzeug==3.0.1